refactor: expose Ktor architecture layers
Split the chapter 12 Ktor example into package-level application, config, routes, service, repository, persistence, and model boundaries so readers see the production shape immediately. Mirror the same boundaries in tests instead of keeping one broad application test, and use Base58.randomString(8) for short H2 database suffixes.

Constraint: The example should stay behaviorally identical while making layer responsibilities obvious from file layout.

Rejected: Keeping all Ktor classes in one file | hides the architecture boundary this example is meant to teach.

Confidence: high

Scope-risk: narrow

Directive: Keep future architecture-focused Ktor examples split by layer once they include more than routes; use Base58.randomString(8) for short resource names.

Tested: ./gradlew --offline :01-ktor-application-architecture:compileKotlin :01-ktor-application-architecture:compileTestKotlin :01-ktor-application-architecture:test; ./gradlew --offline detekt; git diff --check; IntelliJ optimize imports and batch diagnostics; Claude Code refactor review PASS P0=0/P1=0

Not-tested: Fresh IntelliJ editor highlights because the file was not open in the editor; online Gradle resolution because offline verification covered the refactor path.
