+++ 12-production-integration/01-ktor-application-architecture/README.md
@@ -21,12 +21,29 @@ flowchart LR
 ## Learning Goals
 
 - Compose a Ktor application without Spring infrastructure.
+- Read a Ktor codebase through package-level application, route, service,
+  repository, persistence, and model boundaries.
 - Keep routes thin and move validation into the service layer.
 - Keep blocking Exposed JDBC work behind a repository-owned `Dispatchers.IO`
   boundary.
 - Map validation, not-found, malformed JSON, oversized body, and fallback errors
   through Ktor `StatusPages`.
 - Verify routes with Ktor `testApplication`.
+
+## Package Layout
+
+```text
+exposed.examples.ktor.architecture
+├── config       # Ktor plugins, JSON, and HTTP error mapping
+├── model        # Request/response DTOs, commands, records, and validation errors
+├── persistence  # Hikari and Exposed Database lifecycle
+├── repository   # Exposed JDBC repository boundary
+├── routes       # HTTP routes and request body limit handling
+└── service      # Use cases and caller input validation
+```
+
+Tests mirror the same layering: application wiring, route behavior, service
+validation/mapping, and repository persistence/concurrency.
 
 ## Run
