+++ 12-production-integration/01-ktor-application-architecture/src/main/kotlin/exposed/examples/ktor/architecture/config/KtorJson.kt
@@ -1,0 +1,8 @@
+package exposed.examples.ktor.architecture.config
+
+import kotlinx.serialization.json.Json
+
+internal val ApplicationJson = Json {
+    ignoreUnknownKeys = true
+    encodeDefaults = true
+}
